feat: modularises object.patch* and object.new
License: MIT
