+++ js/src/utils/suite.js
@@ -1,16 +1,17 @@
-function createSuite (tests) {
+function createSuite (tests, parent) {
   const suite = (createCommon, options) => {
     Object.keys(tests).forEach(t => {
       const opts = Object.assign({}, options)
+      const suiteName = parent ? `${parent}.${t}` : t
 
       if (Array.isArray(opts.skip)) {
-        if (opts.skip.includes(t)) {
+        if (opts.skip.includes(suiteName)) {
           opts.skip = true
         }
       }
 
       if (Array.isArray(opts.only)) {
-        if (opts.only.includes(t)) {
+        if (opts.only.includes(suiteName)) {
           opts.only = true
         }
       }
